+++ ferrodec-decimal64/src/lib.rs
@@ -28,6 +28,7 @@
 
 #![no_std]
 
+mod bid;
 mod classify_types;
 mod status;